backend/requirements.txt
fastapi
uvicorn
pydantic
duckdb
pandas
pyarrow
python-dotenv
httpx
openpyxl
transformers
torch
accelerate
